Hash the IDs of the notes

+++ client/src/lib/api/note.js
@@ -1,5 +1,6 @@
 import axios from "axios"
 import {PUBLIC_BASE_API_URL} from '$env/static/public'
+import { hashStore } from '../utils/hash';
 
 export const AddNewNote = async (note) => {
     const token = localStorage.getItem('token');
@@ -43,7 +44,9 @@ export const GetNotes = async () => {
     }
 }
 
-export const GetNote = async (id) => {
+export const GetNote = async (hash) => {
+    const id = hashStore.getIDFromHash(hash);
+    if (!id) return null;
     try {
         const res  = await axios.get(`${PUBLIC_BASE_API_URL}/notes/${id}`)
         return res.data
@@ -53,41 +56,54 @@ export const GetNote = async (id) => {
     }
 }
 
-export const DeleteNote = async (id) => {
-    const token = localStorage.getItem('token');
-    
-    if (!token) {
-        return "User not logged in."
-    }
+export const DeleteNote = async (hash) => {
+    const id = hashStore.getIDFromHash(hash);
+    if (!id) return false;
     try {
-        const res  = await axios.delete(`${PUBLIC_BASE_API_URL}/notes/${id}`, {
-            headers: {
-                'Authorization': `Bearer ${token}`,
-                'Content-Type': 'application/json'
-            }
-        })
-        return true
+        const token = localStorage.getItem('token');
+        
+        if (!token) {
+            return "User not logged in."
+        }
+        try {
+            const res  = await axios.delete(`${PUBLIC_BASE_API_URL}/notes/${id}`, {
+                headers: {
+                    'Authorization': `Bearer ${token}`,
+                    'Content-Type': 'application/json'
+                }
+            })
+            return true
+        } catch (error) {
+            console.error(error)
+            alert(error.response?.data.message || "Unable to delete the note.")
+            return false
+        }
     } catch (error) {
         console.error(error)
-        alert(error.response?.data.message || "Unable to delete the note.")
-        return false
     }
 }
 
-export const UpdateNote = async (updatedNote) => {
-    const token = localStorage.getItem('token');
-    
-    if (!token) {
-        return "User not logged in."
-    }
+export const UpdateNote = async (note) => {
+    const id = hashStore.getIDFromHash(note.ID);
+    if (!id) return false;
     try {
-        const res  = await axios.put(`${PUBLIC_BASE_API_URL}/notes/${updatedNote.ID}`, updatedNote, {
-            headers: {
-                'Authorization': `Bearer ${token}`,
-                'Content-Type': 'application/json'
-            }
-        })
-        return res.data
+        const token = localStorage.getItem('token');
+        
+        if (!token) {
+            return "User not logged in."
+        }
+        try {
+            const res  = await axios.put(`${PUBLIC_BASE_API_URL}/notes/${id}`, note, {
+                headers: {
+                    'Authorization': `Bearer ${token}`,
+                    'Content-Type': 'application/json'
+                }
+            })
+            return res.data
+        } catch (error) {
+            console.error(error)
+            alert(error.response?.data.message || "Unable to save changes.")
+        }
     } catch (error) {
         console.error(error)
         alert(error.response?.data.message || "Unable to save changes.")
